--- Presentation/ClientApp/src/environments/environment.development.ts
@@ -1,4 +1,0 @@
-export const environment = {
-    production: false,
-    apiUrl: 'https://localhost:8080/api/',
-};
