+++ Presentation/ClientApp/src/environments/environment.prod.ts
@@ -1,0 +1,4 @@
+export const environment = {
+    production: true,
+    apiUrl: 'https://swift-list-api.onrender.com/api/',
+};
