+++ Presentation/ClientApp/src/environments/environment.ts
@@ -1,4 +1,4 @@
 export const environment = {
-    production: true,
-    apiUrl: 'https://swift-list-api.onrender.com/api/',
+    production: false,
+    apiUrl: 'https://localhost:8080/api/',
 };
